--- src/systems/ec/implementations/react/templates/pages/call-proposal/index.js
@@ -1,1 +1,0 @@
-export default from './src/CallProposalPage';
