+++ src/systems/ec/implementations/react/templates/pages/call-proposals/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "@ecl/ec-react-call-proposal-page",
+  "name": "@ecl/ec-react-call-proposals-page",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.15.0",
@@ -30,7 +30,7 @@
   },
   "devDependencies": {
     "@ecl/ec-layout-grid": "^2.15.0",
-    "@ecl/ec-specs-call-proposal-page": "^2.15.0",
+    "@ecl/ec-specs-call-proposals-page": "^2.15.0",
     "@ecl/ec-utility-spacing": "^2.15.0",
     "@ecl/ec-utility-typography": "^2.15.0"
   },
