+++ src/systems/ec/specs/templates/call-proposals-page/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ecl/ec-specs-call-proposal-page",
+  "name": "@ecl/ec-specs-call-proposals-page",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.15.0",
