+++ src/website/src/pages/ec/content-types/call/proposals/docs/code.mdx
@@ -4,13 +4,13 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-import CallProposalPageExample from '@ecl/ec-react-call-proposal-page/examples/Default';
+import CallProposalsPageExample from '@ecl/ec-react-call-proposals-page/examples/Default';
 
 <Playground
   system="ec"
   selectedKind="templates-pages"
-  selectedStory="call-for-proposal"
+  selectedStory="call-for-proposals"
   hideDemo
 >
-  <CallProposalPageExample />
+  <CallProposalsPageExample />
 </Playground>
